:bug: Fix Targets (Rulesets) with repositories. (#118)
https://issues.redhat.com/browse/MTA-3330

A custom target that references a repository will be created with a
related `RuleSet`. A repository may be defined on the ruleSet. An
analysis task defines _included_ labels which determine which rulesets
are pulled (downloaded) from the hub. Since ruleSets that reference
repositories do not have labels, the addon did not know to download it
and clone the repository.

Solution:
The new rules.ruleSets will be used to pass a list of RuleSet (refs)
selected by the user (by selecting targets). Each will be be
unconditionally fetched by the addon.

For each ruleset with a repository the addon will:
- clone the repository
- update .yaml files - replacing the `labels` element with
`konveyor.io/include=always`.

related: https://github.com/konveyor/tackle2-ui/pull/2034

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>
Signed-off-by: Cherry Picker <noreply@github.com>