:bug: Fix analyzer error relay to task. (#144)
The Issues and Deps report builders need to be returned by the
`Analyzer.Run()` facade having already read the analyzer output files.
This ensures the files are only read once and methods may be invoked in
any order.

This is a regression introduced in 0.6.0.

---

Add hack analyzer script to better support running the addon locally for
testing.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>
Signed-off-by: Cherry Picker <noreply@github.com>