Upload jdtls logs to analysis attachments
In java provider, there is a `.metadata/.log` file with full output of
jdtls. This might be needed for debugging, adding it to analysis tasks
attachments (under more descriptive `provider-jdtls.log` name).

There were konveyor-related entires of this log already part of provider
log, but this PR adds ful content of the jdtls log file.

Fixes: https://github.com/konveyor/tackle2-addon-analyzer/issues/177

Signed-off-by: Marek Aufart <maufart@redhat.com>